[Tizen] packaging: enable xdg-shell with OZONE_WAYLAND_USE_XDG_SHELL
Since this wayland extension is still unstable,
ozone-wayland does now build it but in disabled state by default
so it needs to be enabled on purpose in Tizen's at run time.

To toggle it on, a simple env variable OZONE_WAYLAND_USE_XDG_SHELL
must be defined in user env.
System wide '/etc/profile' mechanism is used
to ensure consistant environment in every login shells.

Note this change relies on
updated version of ozone-wayland to be effective
without it will not cause any arm.

Refer to this page for xdg-shell integration in Tizen :

https://wiki.tizen.org/wiki/Wayland_xdg-shell_protocol