[Tizen] hardcode OZONE_WAYLAND_USE_XDG_SHELL enviroment in service file
This commit force the use of this feature on wayland
(used by tizen's Application.hide() )
in case it's not started from a login shell.

Alternative would be to force a login shell (sh -l -c "...")
and rely on system wide env as introduced
by my other commit about this variable.

But for performance concern it's hardcoded there,
this change may be reverted
if Tizen sessions decide to use login shells then it will work same.

Refer to this page for xdg-shell integration in Tizen :

https://wiki.tizen.org/wiki/Wayland_xdg-shell_protocol